Fix check_no_api_breakages script
Script was failing at "build_and_do" steps. Order seems to matter.
  - When build new then old, build fails.
  - When build old then new, build succeeds.

This might be because "new" build adds files to `.build` that "old" doesn't
expect? Probably because of the `.modulemap` files that were added as
part of the PR. Cleaning up `.build` in between builds seems to fix the
problem.