Add a missing protocol conformance to ClientTransport (#4)
Add a missing protocol conformance to ClientTransport

### Motivation

We lost URLSessionTransport's conformance to ClientTransport along the way somewhere. But that's kind of its main job :)

### Modifications

Re-add the conformance.

### Result

Now the transport can be used as a ClientTransport again.

### Test Plan

Explicitly cast the test transport to be a ClientTransport, rather than using the concrete type.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-urlsession/pull/4