Prepare being compatible with runtime 0.2.0 release (#12)
Prepare being compatible with runtime 0.2.0 release

### Motivation

Runtime 0.2.0 is compatible with 0.1.3+ in the APIs adopted in the transport, so just reflect that in the manifest.

### Modifications

Increased the supported runtime range to include 0.2.x.

### Result

Once released, folks will be able to use 0.1.x of this transport with both 0.1.x and 0.2.x of runtime.

### Test Plan

Will be manually tested once runtime 0.2.0 is released.


Reviewed by: gjcairo

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-urlsession/pull/12