Convert PrivacyInfo.xcprivacy symlinks to files
It is expected behavior of the swift-build build system to copy links as-is, so presenting these as symlinks results in broken symlinks being present in the produced resource bundles, which won't work properly.

On Windows, it even causes the build entirely to fail due to a related bug: https://github.com/apple/swift-foundation/issues/845

Convert these to flat files instead.