Add missing availability guards required when development mode enabled (#337)
### Motivation

When developing code that uses the BoringSSL backend on macOS,
developers set a `development` boolean flag in the Package.swift. We
recently switched from declaring minimum platform versions to annotating
symbols with availability annotations. However, when `development =
true`, it flushes out some annotations that were missing in the first
pass. The result is that `main` currently doesn't build on macOS with
this flag set.

### Modifications

- Add missing availability guards required when development mode
enabled.

### Result:

Code builds macOS with `development = true` in Package.swift.