Fix cmake build.
In #349 we added a new computed property to handle the fact that
ManagedBuffer.capacity is unavailable on OpenBSD. In doing this, however
we encountered some strange errors about this property being `internal
and cannot be referenced from an '@inlinable' function` despite the fact
we marked it `@usableFromInline`. Moving the computed property to the
nested class extension seemed to help, but in recent versions of swift
this caused a different strange compiler error about the computed
property having the `wrong linkage`.

While I don't really understand what is happening here enough to say
whether this is a compiler bug or not, this only seems to be triggering
from `_copyBytes` and `_withVeryUnsafeMutableBytes` still. We want these
to be inlinable, so the next natural workaround seems to be to plumb the
capacity references in by argument instead of referencing `self`. This
makes the callsites a bit wordy, but it seems to work.