Install swiftmodules with full module triple (#354)
Use full module triple in the installed swiftmodules.

### Motivation:

Clean up how the CMake build installs swiftmodules, using the full
module triple instead of just the architecture. This change uses the
compiler to compute the expected module triple instead of attempting to
compute it in CMake script. This is more maintainable and robust to
adding new platforms and architectures.

### Modifications:

Update the CMake build to query the compiler for the full module triple
instead of attempting to compute it in CMake script.