CMake: Declare ASN1 dependency
SwiftCrypto depends on SwiftASN1 which is declared in the exports file,
but does not try to find the project, resulting in build failures of
downstream projects that fail to explicitly call
`find_package(SwiftASN1)` before `find_package(SwiftCrypto)`.

Generally, it is recommended that project config files use
`find_dependency` to search for the dependencies of the project they
represent. If the package is already found, either through
`find_package` in the parent project, or through another means, the
`find_dependency` call is a no-op.