CMake Build: Disable CMP0157 (#387)
There is a bug in CMake CMP0157 where if library A depends on a
non-Swift library B, which in turns depends on a Swift library C,
library A will depend on a swiftmodule for B. B doesn't have a
swiftmodule, so the build fails.

Currently hitting the following error on Windows CI:
`ninja: error: 'swift/CCryptoBoringSSL.swiftmodule', needed by
'lib/libCCryptoBoringSSLShims.lib', missing and no known rule to make
it`

Disabling CMP0157 for now.