Add AES CMAC support
Motivation

AES CMAC is a standardised albeit somewhat uncommon MAC
built on top of the AES block cipher. CMAC pops up in a
few places, mostly where it is convenient to assume that
you have access to AES, but not to a hash function. For
example, AES-SIV is built on CMAC.

Modifications

Add CMAC support
Add tests and test vectors

Result

CMAC is supported